Suggestion: put _JC3D file in it's own subfolder.
Arduino IDE does not like them in one folder (e.g. if one clones the whole repo). That also explaines the "3DsimoKit" subfolder in FW. This is the one the IDE created... The structure would be less prone to errors like:
- FW/3DsimoKit/3DsimoKit.ino
- FW/3DsimoKit_JC3D/3DsimoKit_JC3D.ino
- FW/3DsimoKit_anotherNewFW/3DsimoKit_anotherNewFW.ino    .... and so on...

I wrote a quick tip to the README for anyone else getting an error after downloading the whole repo and trying to use one of the files.